@@ -2,50 +2,50 @@
     "type": "project",
     "license": "proprietary",
     "require": {
-        "php": "^7.1.3",
+        "php": "^7.4",
         "ext-ctype": "*",
         "ext-iconv": "*",
-        "craue/config-bundle": "~2.2",
+        "craue/config-bundle": "~2.4",
         "essence/essence": "^3.5",
-        "fzaninotto/faker": "^1.8",
-        "gos/pubsub-router-bundle": "0.3.2",
-        "gos/web-socket-bundle": "1.9.0-rc3",
+        "fzaninotto/faker": "^1.9",
+        "gos/pubsub-router-bundle": "^2.2",
+        "gos/react-amqp": "^0.3.0",
+        "gos/web-socket-bundle": "^3.0",
         "jbbcode/jbbcode": "^1.4",
-        "oneup/flysystem-bundle": "^3.0",
-        "oneup/uploader-bundle": "^2.1",
-        "react/zmq": "^0.4.0",
-        "sensio/framework-extra-bundle": "^5.2",
-        "sonata-project/admin-bundle": "^3.48",
-        "sonata-project/doctrine-orm-admin-bundle": "^3.9",
-        "stof/doctrine-extensions-bundle": "^1.3",
-        "symfony/asset": "4.2.*",
-        "symfony/console": "4.2.*",
-        "symfony/dom-crawler": "4.2.*",
-        "symfony/dotenv": "4.2.*",
-        "symfony/expression-language": "4.2.*",
-        "symfony/flex": "^1.1",
-        "symfony/form": "4.2.*",
-        "symfony/framework-bundle": "4.2.*",
-        "symfony/monolog-bundle": "^3.1",
+        "oneup/flysystem-bundle": "^3.4",
+        "oneup/uploader-bundle": "^3.0",
+        "stof/doctrine-extensions-bundle": "^1.4",
+        "symfony/asset": "^5.0",
+        "symfony/console": "^5.0",
+        "symfony/dom-crawler": "^5.0",
+        "symfony/dotenv": "^5.0",
+        "symfony/error-handler": "^5.0",
+        "symfony/expression-language": "^5.0",
+        "symfony/flex": "^1.6",
+        "symfony/form": "^5.0",
+        "symfony/framework-bundle": "^5.0",
+        "symfony/monolog-bundle": "^3.5",
         "symfony/orm-pack": "*",
-        "symfony/process": "4.2.*",
-        "symfony/security-bundle": "4.2.*",
+        "symfony/process": "^5.0",
+        "symfony/security-bundle": "^5.0",
         "symfony/serializer-pack": "*",
-        "symfony/swiftmailer-bundle": "^3.1",
-        "symfony/translation": "4.2.*",
-        "symfony/twig-bundle": "4.2.*",
-        "symfony/validator": "4.2.*",
-        "symfony/web-link": "4.2.*",
-        "symfony/webpack-encore-bundle": "^1.1",
-        "symfony/yaml": "4.2.*"
+        "symfony/swiftmailer-bundle": "^3.4",
+        "symfony/templating": "^5.0",
+        "symfony/translation": "^5.0",
+        "symfony/translation-contracts": "^2.0",
+        "symfony/twig-bundle": "^5.0",
+        "symfony/twig-pack": "^1.0",
+        "symfony/validator": "^5.0",
+        "symfony/web-link": "^5.0",
+        "symfony/webpack-encore-bundle": "^1.7",
+        "symfony/yaml": "^5.0"
     },
     "require-dev": {
-        "doctrine/doctrine-fixtures-bundle": "^3.1",
+        "doctrine/doctrine-fixtures-bundle": "^3.3",
         "symfony/debug-pack": "*",
-        "symfony/maker-bundle": "^1.0",
+        "symfony/maker-bundle": "^1.16",
         "symfony/profiler-pack": "^1.0",
-        "symfony/test-pack": "*",
-        "symfony/web-server-bundle": "4.2.*"
+        "symfony/test-pack": "*"
     },
     "config": {
         "preferred-install": {
@@ -89,7 +89,7 @@
     "extra": {
         "symfony": {
             "allow-contrib": false,
-            "require": "4.2.*"
+            "require": "^5.0"
         }
     }
 }
