@@ -6,6 +6,7 @@ use Gos\Bundle\WebSocketBundle\Router\WampRequest;
 use Gos\Bundle\WebSocketBundle\Topic\SecuredTopicInterface;
 use Gos\Bundle\WebSocketBundle\Server\Exception\FirewallRejectionException;
 use Gos\Bundle\WebSocketBundle\Client\ClientManipulator;
+use Gos\Bundle\WebSocketBundle\Client\ClientManipulatorInterface;
 use Ratchet\ConnectionInterface;
 use Ratchet\Wamp\Topic;
 use Ratchet\MessageComponentInterface;
@@ -16,12 +17,20 @@ class ConversationNotificationTopic implements TopicInterface, SecuredTopicInter
 
     private $_usersWhoWriting = array();
 
-    public function __construct(ClientManipulator $clientManipulator)
+    public function __construct(ClientManipulatorInterface $clientManipulator)
     {
         $this->clientManipulator = $clientManipulator;
     }
 
-    public function secure(ConnectionInterface $connection = null, Topic $topic, WampRequest $request, $payload = null, $exclude = null, $eligible = null, $provider = null)
+    public function secure(
+        ?ConnectionInterface $connection,
+        Topic $topic,
+        WampRequest $request,
+        $payload = null,
+        ?array $exclude = [],
+        ?array $eligible = null,
+        ?string $provider = null
+    ): void
     {
         if(!$this->clientManipulator->getClient($connection) instanceof \App\Entity\User)
         {
@@ -98,7 +107,7 @@ class ConversationNotificationTopic implements TopicInterface, SecuredTopicInter
     * Like RPC is will use to prefix the channel
     * @return string
     */
-    public function getName()
+    public function getName() : string
     {
         return 'writing_notification.topic';
     }
